restructuring the file layout to be more clean and easy to navigate
