+++ react-frontend/src/BrickBreakerGame/HelperFunctions/CreateEventListeners.ts
