+++ react-frontend/src/BrickBreakerGame/HelperFunctions/GameTypes.ts
